deployment/requirements.txt
# ISL Recognition - Deployment Dependencies
# ==========================================
# Install with: pip install -r deployment/requirements.txt

# Core ML Framework
tensorflow>=2.12.0

# Data Processing
numpy>=1.23.0
pillow>=9.5.0

# FastAPI Backend
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6

# Streamlit Frontend
streamlit>=1.25.0
plotly>=5.15.0

# Utilities
pandas>=2.0.0
